Add an env meta entry for storing the 'special' contract functions.
These functions should be exposed to the host with expectation of special handling. For now I only introduced the 'check auth' function for the custom account contracts.

While supporting this requires a bit more work than reserving certain function names in host, this approach seems more clean/scalable, as there is no way to have a name collision that leads to unexpected behavior.